Conflicts:
	obdalib-core/src/main/java/org/semanticweb/ontop/mapping/MappingSplitter.java
	obdalib-core/src/main/java/org/semanticweb/ontop/model/impl/OBDADataFactoryImpl.java
	obdalib-core/src/main/java/org/semanticweb/ontop/parser/TurtleOBDAParser.java
	obdalib-core/src/main/java/org/semanticweb/ontop/utils/MetaMappingExpander.java
	obdalib-core/src/test/java/org/semanticweb/ontop/parser/TurtleSyntaxParserTest.java
	obdalib-core/src/test/java/org/semanticweb/ontop/utils/Mapping2DatalogConverterTest.java
	obdalib-owlapi3/src/main/java/org/semanticweb/ontop/owlapi3/directmapping/DirectMappingEngine.java
	obdalib-protege41/pom.xml
	obdalib-protege41/src/main/java/org/semanticweb/ontop/protege4/core/ProtegeOBDAOWLReformulationPlatformFactory.java
	obdalib-protege41/src/main/java/org/semanticweb/ontop/protege4/gui/action/AboxMaterializationAction.java
	obdalib-protege41/src/main/java/org/semanticweb/ontop/protege4/gui/action/R2RMLImportAction.java
	obdalib-r2rml/src/main/java/org/semanticweb/ontop/r2rml/R2RMLReader.java
	obdalib-r2rml/src/main/java/org/semanticweb/ontop/r2rml/R2RMLWriter.java
	ontop-cli/src/main/java/org/semanticweb/ontop/cli/QuestSesameMaterializerCMD.java
	quest-db/src/main/java/org/semanticweb/ontop/owlrefplatform/core/QuestDBConnection.java
	quest-db/src/main/java/org/semanticweb/ontop/owlrefplatform/core/QuestDBStatement.java
	quest-db/src/main/java/org/semanticweb/ontop/owlrefplatform/questdb/QuestDB.java
	quest-db/src/main/java/org/semanticweb/ontop/owlrefplatform/questdb/QuestDBAbstractStore.java
	quest-db/src/main/java/org/semanticweb/ontop/owlrefplatform/questdb/QuestDBClassicStore.java
	quest-db/src/main/java/org/semanticweb/ontop/owlrefplatform/questdb/QuestDBVirtualStore.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWL.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWLCMD.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWLFactory.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWLMaterializerCMD.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWLStatement.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/example/ABoxMaterializerExample.java
	quest-owlapi3/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/example/QuestOWLExample.java
	quest-owlapi3/src/test/java/org/semanticweb/ontop/reformulation/tests/BindTest.java
	quest-owlapi3/src/test/java/org/semanticweb/ontop/reformulation/tests/QuestOWLMaterializerTest.java
	quest-owlapi3/src/test/java/org/semanticweb/ontop/reformulation/tests/SemanticIndexLUBMHTest.java
	quest-sesame/src/test/java/org/semanticweb/ontop/sql/TestQuestImplicitDBConstraints.java
	quest-test/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWLExample_OntowisTests.java
	quest-test/src/main/java/org/semanticweb/ontop/owlrefplatform/owlapi3/QuestOWLExample_ReasoningDisabled.java
	quest-test/src/main/java/org/semanticweb/ontop/quest/sparql/OracleRegexpTest.java
	quest-test/src/main/java/org/semanticweb/ontop/quest/sparql/OracleRegexpTestSpace.java
	quest-test/src/test/java/org/semanticweb/ontop/ask/H2ASKTest.java
	quest-test/src/test/java/org/semanticweb/ontop/obda/ImdbTestPostgres.java
	quest-test/src/test/java/org/semanticweb/ontop/obda/OntologyTypesTest.java
	quest-test/src/test/java/org/semanticweb/ontop/obda/QuestOWLExampleNotLatin.java
	quest-test/src/test/java/org/semanticweb/ontop/obda/R2rmlCheckerTest.java
	quest-test/src/test/java/org/semanticweb/ontop/sesame/repository/test/SesameMaterializerCmdTest.java
	quest-test/src/test/java/org/semanticweb/ontop/sesame/tests/general/InconsistencyCheckingVirtualTest.java
	quest-test/src/test/java/org/semanticweb/ontop/unfold/LeftJoinTest1Virtual.java
	quest-test/src/test/java/org/semanticweb/ontop/unfold/LeftJoinTest2Virtual.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/Quest.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/QuestConnection.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/QuestPreferences.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/QuestStatement.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/QuestUnfolder.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/abox/QuestMaterializer.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/abox/RDBMSDataRepositoryManager.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/abox/RDBMSSIRepositoryManager.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/basicoperations/DatalogNormalizer.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/dagjgrapht/NamedDAG.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/mappingprocessing/MappingDataTypeRepair.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/queryevaluation/SQLAdapterFactory.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/sql/SQLGenerator.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/srcquerygeneration/NativeQueryGenerator.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/translator/SQLMappingVocabularyFixer.java
	reformulation-core/src/main/java/org/semanticweb/ontop/owlrefplatform/core/translator/SparqlAlgebraToDatalogTranslator.java
	reformulation-core/src/test/java/org/semanticweb/ontop/owlrefplatform/core/abox/VirtualABoxMaterializerTest.java

+++ obdalib-protege41/src/main/java/org/semanticweb/ontop/protege4/views/QueryInterfaceView.java
@@ -148,7 +148,7 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 		setupListeners();
 		
 		// Setting up actions for all the buttons of this view.
-		resultTablePanel.setCountAllTuplesActionForUCQ(new OBDADataQueryAction<Integer>("Counting tuples...", QueryInterfaceView.this) {
+		resultTablePanel.setCountAllTuplesActionForUCQ(new OBDADataQueryAction<Long>("Counting tuples...", QueryInterfaceView.this) {
 			@Override
 			public OWLEditorKit getEditorKit(){
 				return getOWLEditorKit();
@@ -158,13 +158,13 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 				return -1;
 			}
 			@Override
-			public void handleResult(Integer result){
+			public void handleResult(Long result){
 				updateTablePanelStatus(result);	
 			}
 
 
 			@Override
-			public Integer executeQuery(QuestOWLStatement st, String query) throws OWLException {
+			public Long executeQuery(QuestOWLStatement st, String query) throws OWLException {
 				return st.getTupleCount(query);
 			}
 
@@ -359,8 +359,8 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 
 	
 	private class QueryStatusUpdater implements Runnable{
-		int result;
-		QueryStatusUpdater(int result){
+		long result;
+		QueryStatusUpdater(long result){
 			this.result = result;
 		}
 		
@@ -368,7 +368,7 @@ public class QueryInterfaceView extends AbstractOWLViewComponent implements Save
 			queryEditorPanel.updateStatus(result);
 		}
 	}
-	protected void updateTablePanelStatus(int result) {
+	protected void updateTablePanelStatus(Long result) {
 		if (result != -1) {
 			Runnable status_updater = new QueryStatusUpdater(result);
 			SwingUtilities.invokeLater(status_updater);
